fixed typo. Regarding PR #40

+++ mp3/decode.go
@@ -90,7 +90,7 @@ func (d *decoder) Seek(p int) error {
 	if err != nil {
 		return errors.Wrap(err, "mp3")
 	}
-	d.pos = int64(p) * gomp3BytesPerFrame
+	d.pos = p * gomp3BytesPerFrame
 	return nil
 }
